% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neg_SPH_HexCer.R
\name{create_neg_HexCer}
\alias{create_neg_HexCer}
\title{Generation of fragmentation spectra for PC \link{M+FA-H}-}
\usage{
create_neg_HexCer(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of fragmentation spectra for PC \link{M+FA-H}-
}
