% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neg_GP_PA.R
\name{create_neg_PA}
\alias{create_neg_PA}
\title{Generation of fragmentation spectra for PA \link{M-H}-}
\usage{
create_neg_PA(lipid_info, adduct = "[M-H]-", template = NA, ...)
}
\description{
Generation of fragmentation spectra for PA \link{M-H}-
}
