% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neg_PC.R
\name{create_neg_PC}
\alias{create_neg_PC}
\title{Generation of fragmentation spectra for PC [M+FA-H]-}
\usage{
create_neg_PC(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of fragmentation spectra for PC [M+FA-H]-
}
