% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neg_PC.R
\name{create_neg_PC_MHAcH}
\alias{create_neg_PC_MHAcH}
\title{Generation of fragmentation spectra for PC [M+HAc-H]-}
\usage{
create_neg_PC_MHAcH(lipid_info, template = NA, ...)
}
\description{
Generation of fragmentation spectra for PC [M+HAc-H]-
}
