% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neg_GP_PC-O.R
\name{create_neg_PCO}
\alias{create_neg_PCO}
\title{Generation of fragmentation spectra for PC \link{M+FA-H}-}
\usage{
create_neg_PCO(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of fragmentation spectra for PC \link{M+FA-H}-
}
