% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neg_GP_PE.R
\name{create_neg_PE}
\alias{create_neg_PE}
\title{Generation of fragmentation spectra for PE \link{M-H}-}
\usage{
create_neg_PE(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of fragmentation spectra for PE \link{M-H}-
}
