% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neg_GP_PE-O.R
\name{create_neg_PEO}
\alias{create_neg_PEO}
\title{Generation of fragmentation spectra for PEO}
\usage{
create_neg_PEO(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of fragmentation spectra for PEO
}
