% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neg_GP_PE-P.R
\name{create_neg_PEP}
\alias{create_neg_PEP}
\title{Generation of fragmentation spectra for PEP \link{M-H}-}
\usage{
create_neg_PEP(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of fragmentation spectra for PEP \link{M-H}-
}
