% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neg_PG.R
\name{create_neg_PG}
\alias{create_neg_PG}
\title{Generation of fragmentation spectra for PE [M-H]-}
\usage{
create_neg_PG(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of fragmentation spectra for PE [M-H]-
}
