% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_neg_PS.R
\name{create_neg_PS_MH}
\alias{create_neg_PS_MH}
\title{Generation of fragmentation spectra for PE [M-H]-}
\usage{
create_neg_PS_MH(lipid_info, template = NA, ...)
}
\description{
Generation of fragmentation spectra for PE [M-H]-
}
