% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_SPH_HexCer.R
\name{create_pos_HexCer}
\alias{create_pos_HexCer}
\title{Generation of Cer MS2 spectrum}
\usage{
create_pos_HexCer(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of Cer MS2 spectrum
}
