% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_GP_LPC.R
\name{create_pos_LPC}
\alias{create_pos_LPC}
\title{Generation of LPC \link{M+H}+ MS2 spectrum}
\usage{
create_pos_LPC(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of LPC \link{M+H}+ MS2 spectrum
}
