% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_GL_MG.R
\name{create_pos_MG}
\alias{create_pos_MG}
\title{Generation of PC \link{M+H}+ MS2 spectrum}
\usage{
create_pos_MG(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of PC \link{M+H}+ MS2 spectrum
}
