% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_GP_PA.R
\name{create_pos_PA}
\alias{create_pos_PA}
\title{Generation of PA \link{M+H}+ MS2 spectrum}
\usage{
create_pos_PA(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of PA \link{M+H}+ MS2 spectrum
}
