% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_PC.R
\name{create_pos_PC_MNa}
\alias{create_pos_PC_MNa}
\title{Generation of PC [M+H]+ MS2 spectrum}
\usage{
create_pos_PC_MNa(lipid_info, template = NA, ...)
}
\description{
Generation of PC [M+H]+ MS2 spectrum
}
