% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_PE.R
\name{create_pos_PE_MH}
\alias{create_pos_PE_MH}
\title{Generation of PC [M+H]+ MS2 spectrum}
\usage{
create_pos_PE_MH(lipid_info, template = NA, ...)
}
\description{
Generation of PC [M+H]+ MS2 spectrum
}
