% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_PE.R
\name{create_pos_PE_MNa}
\alias{create_pos_PE_MNa}
\title{Generation of PC [M+Na]+ MS2 spectrum}
\usage{
create_pos_PE_MNa(lipid_info, template = NA, ...)
}
\description{
Generation of PC [M+Na]+ MS2 spectrum
}
