% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_GP_PI.R
\name{create_pos_PI}
\alias{create_pos_PI}
\title{Generation of PC \link{M+H}+ MS2 spectrum}
\usage{
create_pos_PI(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of PC \link{M+H}+ MS2 spectrum
}
