% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_PS.R
\name{create_pos_PS_MH}
\alias{create_pos_PS_MH}
\title{Generation of PS [M+H]+ MS2 spectrum}
\usage{
create_pos_PS_MH(lipid_info, template = NA, ...)
}
\description{
Generation of PS [M+H]+ MS2 spectrum
}
