% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_SPH_SM.R
\name{create_pos_SM}
\alias{create_pos_SM}
\title{Generation of PC \link{M+H}+ MS2 spectrum}
\usage{
create_pos_SM(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of PC \link{M+H}+ MS2 spectrum
}
