% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pos_GL_TG.R
\name{create_pos_TG}
\alias{create_pos_TG}
\title{Generation of TG \link{M+Na}+ or \link{M+NH4}+ MS2 spectrum}
\usage{
create_pos_TG(lipid_info, adduct, template = NA, ...)
}
\description{
Generation of TG \link{M+Na}+ or \link{M+NH4}+ MS2 spectrum
}
